creation classe abstraite combattant, la classe personnage hérite de combattant, la classe monstre hérite de combattant
