+++ src/main/java/fr/adsa/Main.java
@@ -7,6 +7,6 @@ public class Main {
     public static void main(String[] args) {
         Monde m = new Monde();
         Personnage hero = m.createPersonnage();
-        m.afficherInfosPersonnage(hero);
+        m.afficherInfos(hero);
     }
 }
