+++ src/main/java/fr/adsa/model/Monstre.java
@@ -1,0 +1,9 @@
+package fr.adsa.model;
+
+import fr.adsa.abstrait.Combattant;
+
+public class Monstre extends Combattant {
+    public Monstre(String nom, int pdv, int degats) {
+        super(nom, pdv, degats);
+    }
+}
